+++ src/popup/panels/WhatsNewPanel.vue
@@ -2,19 +2,14 @@
   <div>
     <h2>What's new</h2>
     <p>Full changelog for older versions <a href="https://github.com/tamius-han/ultrawidify/blob/master/CHANGELOG.md">is available here</a>.</p>
-    <p class="label">5.0.1</p>
+    <p class="label">5.0.4</p>
     <ul>
       <li>
-        Disabled Chrome's and Edge's zoom limitation.
-      </li>
-      <li>
-        Added an option to toggle and further configure zoom limitation.
+        Updated config for disney+. Hopefully it works now. Special thanks to <a href="https://github.com/tamius-han/ultrawidify/issues/84#issuecomment-846334005" target="_blank">@jwannebo</a>.
       </li>
     </ul>
     <p>
-      <small><b>NOTE:</b> zoom limitations were introduced as a workaround for a bug caused by Chrome's/Edge's faulty hardware acceleration
-      which caused videos to be incorrectly stretched on nVidia GPUs while in fullscreens. While this issue is reportedly fixed as
-      of few weeks ago, a few people reported having this issue within the past two weeks.</small>
+      <small><b>NOTE from older versions:</b> zoom limitations were introduced as a workaround for a bug caused by Chrome's/Edge's faulty hardware acceleration.</small>
     </p>
     <p>
       <small>If you experience issues with videos being stretched incorrectly at certain zoom levels, go to:</small><br/>